% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Computes The Shortest Path Between Two Given Nodes}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm/}
}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Parameter 1}

\item{init_node}{Parameter 2}
}
\value{
Returns the distance between two given nodes.
}
\description{
It is an algorithm for finding the shortest paths between nodes in a graph.
}
