% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Computes The Greatest Common Divisor Of Two Numbers(Integers).}
\source{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{One of the integers of the two required to calculate the GCD}

\item{y}{One of the integers of the two required to calculate the GCD}
}
\value{
This code returns the greatest common divisor
}
\description{
Computes the greatest common divisor for two given numbers(integers)
in an efficient method.
}
